+++ HotelTamagotchi.Web/Controllers/TamagotchiController.cs
@@ -26,12 +26,22 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi
         public ActionResult Index()
         {
-            return View(TamagotchiRepo.GetAll());
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            return View(TamagotchiRepo.GetAllFromUser((int)Session["UserId"]));
         }
 
         // GET: Tamagotchi/Details/5
         public ActionResult Details(int? id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -47,6 +57,16 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Create
         public ActionResult Create()
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            else if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             return View();
         }
 
@@ -57,6 +77,18 @@ namespace HotelTamagotchi.Web.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,HotelRoomId,Name,Age,Pennies,Level,Health,Boredom,Alive")] TamagotchiViewModel tamagotchi)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            else if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
+
+            tamagotchi.UserId = (int)Session["UserId"];
             if (ModelState.IsValid)
             {
                 TamagotchiRepo.Add(tamagotchi);
@@ -69,6 +101,16 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer) || !Session["User"].Equals(TamagotchiRepo.Find(id).User.Username))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Tamagotchi");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -86,8 +128,18 @@ namespace HotelTamagotchi.Web.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,HotelRoomId,Name,Age,Pennies,Level,Health,Boredom,Alive")] TamagotchiViewModel tamagotchi)
+        public ActionResult Edit(TamagotchiViewModel tamagotchi)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (ModelState.IsValid)
             {
                 TamagotchiRepo.SetChanged(tamagotchi);
@@ -99,6 +151,16 @@ namespace HotelTamagotchi.Web.Controllers
         // GET: Tamagotchi/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer) || !Session["User"].Equals(TamagotchiRepo.Find(id).User.Username))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -116,6 +178,16 @@ namespace HotelTamagotchi.Web.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (Session["User"] == null)
+            {
+                TempData["NotLoggedIn"] = "Please login or register to continue!";
+                return RedirectToAction("Index", "Home");
+            }
+            if (!Session["Role"].Equals(UserRole.Customer) || !Session["User"].Equals(TamagotchiRepo.Find(id).User.Username))
+            {
+                TempData["NotAuthorized"] = "You are not authorized to perform these actions!";
+                return RedirectToAction("Index", "Home");
+            }
             TamagotchiViewModel tamagotchi = TamagotchiRepo.Find(id);
             TamagotchiRepo.Remove(tamagotchi);
             return RedirectToAction("Index");
